Extract and move option types
