+++ src/lib/readable-stream/iterator-options.ts
@@ -1,0 +1,3 @@
+export interface ReadableStreamIteratorOptions {
+  preventCancel?: boolean;
+}
