+++ src/lib/readable-stream/pipe-options.ts
@@ -1,0 +1,8 @@
+import { AbortSignal } from '../abort-signal';
+
+export interface PipeOptions {
+  preventAbort?: boolean;
+  preventCancel?: boolean;
+  preventClose?: boolean;
+  signal?: AbortSignal;
+}
